==> docs/02_Gannon_Corruption_TrumpTown.Rmd <==
---
title: "02 Justice Department"
author: "J Andres Gannon, Danil Gromakov, Katelyn Hess, and Yulia Krylova"
date: "May 2020"
output:
  html_document:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(magrittr)
library(ggplot2)
```

This document loads, cleans, and processes data from ProPublica's Trump Tower's archives concerning Department of Justice appointments during the Trump administration. We collect data on their employment history, education history, and membership in various organizations to identify patterns in how these Justice Department officials were chosen.

# Base data
## Load
```{r message = FALSE}
df <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1hPOBOyTujUReQzpjZ9PqIhj23V5C74EygMnmXIi6BgI/edit?usp=sharing",
                                    sheet = "doj_list")

df <- janitor::clean_names(df)
```

## Summarize basic info
```{r, fig.width = 16, fig.height = 16, warning = FALSE}
DT::datatable(df)
```

### Wikidata SPARQL query
```{r, message = FALSE}
df_wiki <- read.csv(file = paste0(here::here(), '/data/query_doj.csv'))
```

Extract wikidata IDs from SPARQL query for easier merging
```{r}
df_wiki$id <- sub(".*entity/", "", df_wiki$item)
```

Subset wiki query to qualities of interest and do minor cleaning
```{r}
# Subset
df_wiki <- df_wiki %>% dplyr::select(c("id", "itemLabel", "sex_or_genderLabel", "date_of_birth", "place_of_birthLabel", "educated_atLabel"))
df_wiki <- unique(df_wiki)

# Fix DoB type
df_wiki$date_of_birth <- as.Date(df_wiki$date_of_birth)

# Add age var
age <- eeptools::age_calc(na.omit(df_wiki$date_of_birth), units = "years")
df_wiki$age <- NA
df_wiki$age[!is.na(df_wiki$date_of_birth)] <- age
df_wiki$age <- round(df_wiki$age)

# Fix Sessions errors. 2 birthplaces with differing primary education
df_wiki <- df_wiki[- grep("Hybart, Alabama", df_wiki$place_of_birthLabel), ]
df_wiki <- df_wiki[- grep("primary school", df_wiki$educated_atLabel), ]

# educated_at var should be in multiple columns so there's one row per prosecutor
df_wiki <- df_wiki %>% dplyr::group_by(id) %>%
       dplyr::mutate(rn = paste0("educated_atLabel", dplyr::row_number())) %>%
       tidyr::spread(rn, educated_atLabel)

# Examine
DT::datatable(df_wiki)
```

## Initial stats
### Sex or gender
```{r, fig.width=16, fig.height=10}
ggplot(df_wiki, aes(sex_or_genderLabel, ..count..)) + 
  geom_bar() + 
  labs(title = "Department of Justice Official", x = "Sex", y = "Count", fill = "") + 
  geom_text(stat = 'count', aes(label = ..count..), vjust = -1, size = 6) +
  theme(plot.title = element_text(hjust = 0.5),panel.grid = element_blank(),
        text = element_text(size = 24))
```

### Age
```{r}
ggplot(df_wiki, aes(age, ..count..)) + 
  geom_histogram() + 
  labs(title = "Department of Justice officials", x = "Age", y = "Count", fill = "") + 
  theme(plot.title = element_text(hjust = 0.5),panel.grid = element_blank(),
        text = element_text(size = 24))
```

## Merge wikidata with base
```{r, warning = FALSE}
df_full <- dplyr::left_join(df, df_wiki, by = c('wikidata_id' = 'id'))

# Only keep non-redundant rows
df_full <- df_full %>% dplyr::select(c("attorney", "sex_or_genderLabel", "date_of_birth", "place_of_birthLabel", "age")) %>%
  dplyr::rename(sex = sex_or_genderLabel, birthplace = place_of_birthLabel)
```

# Employment
## Googlesheet
### Load
We have 2 datasets for justice department employment info. The first is the google doc we have been manually curating and the second is the trump town employment data

```{r}
work <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1hPOBOyTujUReQzpjZ9PqIhj23V5C74EygMnmXIi6BgI/edit?usp=sharing",
                                    sheet = "doj_employ")

work <- work %>% dplyr::distinct()
work <- janitor::clean_names(work)
```

### Clean
```{r}
# Check employer strings
sort(unique(work$employer))

# Fix start and end date
work$start_date[work$start_date == "Unknown"] <- NA
work$end_date[work$end_date == "Unknown"] <- NA
```

## Trump town
### Load
```{r}
df_staff <- read.csv(file = paste0(here::here(), '/network_WH_outOrganisations/trumptown_data-store_1019/trumptown_staffer_organizations.csv'))
df_staff <- janitor::clean_names(df_staff)
```

### Clean
```{r}
# Subset to only attorneys
df_staff <- df_staff %>% dplyr::filter(stringr::str_detect(position_title_1, stringr::regex("attorney", ignore_case = TRUE)) | stringr::str_detect(position_title_1, stringr::regex("solicitor", ignore_case = TRUE)))

df_staff <- df_staff %>% dplyr::select(agency_name, position_title_1, staffer_name, name_organization, start_date, end_date, role_description) %>%
  dplyr::rename(position_title = position_title_1, attorney = staffer_name, employer = name_organization, title = role_description) %>%
  dplyr::distinct()

df_staff$start_date <- format(as.Date(df_staff$start_date, format = "%Y-%m-%d"), "%Y")
df_staff$start_date <- as.numeric(df_staff$start_date)

df_staff$end_date <- format(as.Date(df_staff$end_date, format = "%Y-%m-%d"), "%Y")
df_staff$end_date <- as.numeric(df_staff$end_date)

df_staff$notes <- NA
```

### Get list of senate appts
```{r}
df_merge <- df_staff %>% dplyr::select(agency_name, position_title, attorney) %>%
  dplyr::distinct()

df_merge$senate_appt <- NA

df_merge$senate_appt[df_merge$attorney == "Beth Ann Williams" |
                      df_merge$attorney == "Brian Allen Benczkowski" |
                      df_merge$attorney == "Eric S. Dreiband" |
                      df_merge$attorney == "Jefferson B. Sessions" |
                      df_merge$attorney == "Jeffrey Bossert Clark" |
                      df_merge$attorney == "John C. Demers" |
                      df_merge$attorney == "Joseph H. Hunt" |
                      df_merge$attorney == "Noel Francisco" |
                      df_merge$attorney == "Rachel L. Brand" |
                      df_merge$attorney == "Rod J. Rosenstein" |
                      df_merge$attorney == "Steven A. Engel" |
                      df_merge$attorney == "William P. Barr"] <- 1

df_merge <- df_merge %>% dplyr::filter(senate_appt == 1)

df_merge$senate_appt <- NULL
```

## Merge
Get a single list of all attorneys
```{r}
# Master list of attorneys
attorneys <- as.data.frame(unique(df_merge$attorney)) %>%
  dplyr::rename(attorney = "unique(df_merge$attorney)")
attorneys2 <- as.data.frame(unique(work$attorney)) %>%
  dplyr::rename(attorney = "unique(work$attorney)")

attorney_list <- dplyr::full_join(attorneys, attorneys2) %>%
  dplyr::distinct()
attorney_list$keep <- 1

# Merge into each original df
df_pen <- dplyr::left_join(df_staff, attorney_list) %>%
  dplyr::filter(keep == 1) %>%
  dplyr::select(-c("keep", "notes"))

work_pen <- dplyr::left_join(work, attorney_list) %>%
  dplyr::filter(keep == 1) %>%
  dplyr::select(-c("keep"))

df_full <- dplyr::bind_rows(df_pen, work_pen) %>%
  dplyr::distinct()

# Fix string matches
sort(unique(df_full$employer))

df_full$employer[agrep("Kirkland & Ellis", df_full$employer)] <- "Kirkland & Ellis"
df_full$employer[df_full$employer == "A. S. and J.S. Irrevocable Trust"] <- "A. S. Trust"
df_full$employer[df_full$employer == "BP America (client of Kirkland & Ellis LLP)"] <- "BP America"
df_full$employer[df_full$employer == "Brownstein Hyatt Farber Schreck, LLP"] <- "Brownstein Hyatt Farber Schreck LLP"
df_full$employer[df_full$employer == "Caesars Entertainment Operating Company (client of Kirkland & Ellis LLP)"] <- "Caesars Interactive Entertainment, Inc."
df_full$employer[df_full$employer == "Federalist Society for Law & Policy"] <- "Federalist Society"
df_full$employer[df_full$employer == "Georgetown Law Center"] <- "Georgetown University Law Center"
df_full$employer[df_full$employer == "Jones Day"] <- "Jones Day LLP"
df_full$employer[df_full$employer == "Pepperdine University"] <- "Pepperdine University School of Law"
df_full$employer[df_full$employer == "The Boeing Company"] <- "Boeing Company"
df_full$employer[df_full$employer == "University of North Carolina, Board of Trustees of UNC, W. Louis Bissette, Jr., Margaret Spellings"] <- "University of North Carolina"
df_full$employer[df_full$employer == "Verizon Communication Inc"] <- "Verizon Communications"
df_full$employer[df_full$employer == "Volkswagen Group of America (client of Kirkland & Ellis LLP)"] <- "Volkswagen Group of America"

sort(unique(df_full$employer))

df_full <- df_full %>% dplyr::distinct()
```

## Prep for network
```{r}
# Places of employment with the most unique attorneys
df_count <- df_full %>% dplyr::select(attorney, employer) %>%
  dplyr::distinct() %>%
  dplyr::group_by(employer) %>%
  dplyr::summarise(count = dplyr::n()) %>%
  dplyr::filter(count > 1)

DT::datatable(df_count)
```

## Save
````{r}
work <- df_full

work <- write.csv(work, paste0(here::here(), "/data/", "justicedept_employment.csv"))
```

# School
## Load
```{r}
edu <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1hPOBOyTujUReQzpjZ9PqIhj23V5C74EygMnmXIi6BgI/edit?usp=sharing",
                                   sheet = "doj_edu",
                                  col_types = "ccccnnnnc")

edu <- janitor::clean_names(edu)
```

## Clean
```{r}
## Collapse degrees
sort(unique(edu$degree))

edu$degree[edu$degree == "AA" |
             edu$degree == "AB" |
             edu$degree == "BA" |
             edu$degree == "BBA" |
             edu$degree == "BS"] <- "bachelors"
 
edu$degree[edu$degree == "LLM" |
             edu$degree == "MA" |
             edu$degree == "MBA" |
             edu$degree == "MMgt" |
             edu$degree == "MPA" |
             edu$degree == "MPhil" |
             edu$degree == "MS"] <- "masters"

edu$degree[edu$degree == "PhD" |
             edu$degree == "EdD"] <- "doctorate"

edu$degree[edu$degree == "JD" |
             edu$degree == "LLB"] <- "law"

sort(unique(edu$degree))
```

## Prep for network
```{r}
# Make wide
edu_clean <- edu %>% dplyr::select(item_label, educated_at_label, degree) %>%
  tidyr::spread(degree, educated_at_label) %>%
  dplyr::select(item_label, bachelors, masters, law, doctorate)
edu_clean$law <- paste(edu_clean$law, "law", sep = "_")

# Merge, but careful to split undergrad and law
edu_merge <- edu %>% tidyr::unite(educated_combo, educated_at_label, degree, sep = "_", remove = TRUE) %>%
  dplyr::select(educated_combo, law_rank, law_conservative_score) %>%
  dplyr::distinct()

edu_all <- dplyr::left_join(edu_clean, edu_merge, by = c("law" = "educated_combo")) %>%
  dplyr::distinct()
edu_all$law <- gsub("_law", "", paste(edu_all$law))
edu_all$law[edu_all$law == "NA"] <- NA

# Clean each school col
sort(unique(edu_all$bachelors))
sort(unique(edu_all$masters))
sort(unique(edu_all$law))
```

## Results
```{r}
DT::datatable(edu_all)

GGally::ggpairs(edu_all[ , c('law_rank', 'law_conservative_score')], columnLabels = c("Law School Rank", "Law School Conservatism"))
```

## Save
```{r}
edu <- write.csv(edu_all, paste0(here::here(), "/data/", "justicedept_edu.csv"))
```

# Memberships
## Load
```{r}
memb <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1hPOBOyTujUReQzpjZ9PqIhj23V5C74EygMnmXIi6BgI/edit?usp=sharing",
                                    sheet = "doj_memb",
                                  col_types = "nccnnc")

memb <- janitor::clean_names(memb)
```

## Prep for network
```{r}
# Make wide
memb_clean <- memb %>% dplyr::filter(senate_appt == 1) %>%
  dplyr::select(-c("political_affiliation", "notes"))
```

## Results
```{r}
DT::datatable(memb_clean)

memb_upset <- memb_clean %>%
  dplyr::rename("Federalist Society" = federalist_society, "Heritage Foundation" = heritage_foundation)

memb_upset <- as.data.frame(sapply(memb_upset, as.numeric))
memb_upset[is.na(memb_upset)] <- 0

UpSetR::upset(memb_upset,
      number.angles = 0, point.size = 3.5, line.size = 2, text.scale = 2,
      mainbar.y.label = "Number of Prosecutors", order.by = "freq")
```

## Save
```{r}
memb <- write.csv(memb_clean, paste0(here::here(), "/data/", "justicedept_membership.csv"))
```

# Final
## Load all cleaned
```{r}
# Load
employ <- read.csv(paste0(here::here(), "/data/", "justicedept_employment.csv"))
edu <- read.csv(paste0(here::here(), "/data/", "justicedept_edu.csv"))
memb <- read.csv(paste0(here::here(), "/data/", "justicedept_membership.csv"))
```

## Clean
```{r}
colnames(employ)
employ <- employ %>% dplyr::select(-c("X")) %>%
  dplyr::rename(person = attorney, agency_title = position_title)

colnames(edu)
edu <- edu %>% dplyr::select(-c("X")) %>%
  dplyr::rename(person = item_label)

colnames(memb)
memb <- memb %>% dplyr::select(-c("X", "senate_appt")) %>%
  dplyr::rename(person = item_label)
```

## Merge
Since the only one with more columns is employment, we merge everything into that
```{r}
full <- dplyr::left_join(employ, edu)
full <- dplyr::left_join(full, memb)

full <- full %>% dplyr::distinct()
```

## Save
```{r}
write.csv(full, paste0(here::here(), "/data/", "justicedept_final.csv"))
```
