==> docs/03_Gannon_Corruption_All.Rmd <==
---
title: "03 Combination"
author: "J Andres Gannon, Danil Gromakov, Katelyn Hess, and Yulia Krylova"
date: "May 2020"
output:
  html_document:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(magrittr)
library(ggplot2)
```

This document produces combines the new data collected regarding US prosecutor and Department of Justice officials appointed during the Trump administration

# Load data
```{r}
prosecutors <- read.csv(file = paste0(here::here(), '/data/prosecutors_final.csv'))
justicedept <- read.csv(file = paste0(here::here(), '/data/justicedept_final.csv'))
```

Combine into a single df
```{r}
# Check column names and fix for merging
sort(colnames(prosecutors))
sort(colnames(justicedept))

prosecutors <- prosecutors %>% dplyr::rename(american_center_for_law_and_justice = american_center_for_law_and_justice_tie,
                                             christian_legal_society_bar_association = christian_legal_society_bar_association_tie,
                                             position_current = district,
                                             employer_past = employer,
                                             federalist_society = federalist_society_tie,
                                             heritage_foundation = heritage_foundation_tie) %>%
  dplyr::select(-c("X"))
prosecutors$source <- "prosecutor"

justicedept <- justicedept %>% dplyr::rename(position_current = agency_title,
                                             employer_past = employer) %>%
                dplyr::select(-c("X", "title", "agency_name"))
justicedept$source <- "justicedept"

df <- dplyr::bind_rows(prosecutors, justicedept) %>%
  dplyr::select("person", "position_current", "source",
                "date_of_appointment",
                "date_of_birth", "age", "birthplace", "sex",
                "bachelors", "masters", "doctorate", "law", "law_conservative_score", "law_rank",
                "barr_appointment", "political_affiliation", 
                "employer_past", "firm_conservative_score", "start_date", "end_date",
                "american_center_for_law_and_justice", "christian_legal_society_bar_association", "federalist_society", "heritage_foundation", "national_rifle_association_tie", "republican_national_lawyers_association_tie")

df_noemploy <- df %>% dplyr::select(-c("employer_past", "firm_conservative_score", "start_date", "end_date")) %>%
  dplyr::distinct()
```

# Aggregate
Here, we use all the covariates gathered above to identify whether each person has a connection to Trump, Trump's allies, a conservative agenda, a career professional, or other

## Load
This df has the final aggregated list of unique persons from both datasets we have created. We aggregate each of the main variables of interest to different levels to help us tell different stories.
```{r, eval}
# core <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1hPOBOyTujUReQzpjZ9PqIhj23V5C74EygMnmXIi6BgI/edit?usp=sharing",
#                                    sheet = "all")

core <- read.csv(file = paste0(here::here(), '/data/raw/TrumpProsecutors - all.csv'))

core <- janitor::clean_names(core)
```

## Conservative agenda
We code whether someone has a conservative agenda based on whether they attended a conservative law school, worked at a conservative law firm, have tied to conservative organizations, or are a conservative/Republican elected politician (state senator or state representative). We create a composite index coded as 1 if any of these criteria are met, but also maintain each individual criteria to analyze separately

### Law school
Using data on ideological leaning of law schools, we identify what people attended conservative leaning law schools. Since the ideological bias of law school is generally liberal, we scale it account for relative conservatism
```{r}
# Calculate average law school conservatism
school <- read.csv(file = paste0(here::here(), '/data/ideology_lawschool.csv'))
school$score_scaled <- scale(school$conservatism)

# Merge scaled score into a temp df
school_temp <- df_noemploy %>% dplyr::select(person, law, law_conservative_score) %>%
  dplyr::rename(school = law) %>%
  dplyr::distinct()

school_temp <- dplyr::left_join(school_temp, school)

# Create new column for worked at a more conservative than average firm
school_temp <- school_temp %>% dplyr::mutate(conservative = dplyr::if_else(score_scaled > 0, 1, 0)) %>%
  dplyr::select(person, conservative) %>%
  dplyr::rename(connection_conservative_lawschool = conservative) %>%
  dplyr::distinct()

# Merge that into core
tomerge <- school_temp
```

### Law firm
We repeat the process for law firms, calculating their conservative bias and then identifying people that worked at relatively conservative law firms
```{r}
# Calculate average law firm conservatism
firm <- read.csv(file = paste0(here::here(), '/data/ideology_lawfirm.csv'))
firm$score_scaled <- scale(firm$conservatism)

# Merge scaled score into a temp df
firm_temp <- df %>% dplyr::select(person, employer_past, firm_conservative_score) %>%
  dplyr::rename(firm = employer_past) %>%
  dplyr::distinct()

# Fix names for string matching
firm_temp$firm <- gsub(" LLP", "", firm_temp$firm)

firm_temp <- dplyr::left_join(firm_temp, firm)

# Create new column for worked at a more conservative than average firm
firm_temp <- firm_temp %>% dplyr::mutate(conservative = dplyr::if_else(score_scaled > 0, 1, 0)) %>%
  dplyr::select(person, conservative) %>%
  dplyr::rename(connection_conservative_firm = conservative)

firm_temp$connection_conservative_firm[is.na(firm_temp$connection_conservative_firm)] <- 0

firm_temp <- firm_temp %>% dplyr::distinct()

firm_temp <- firm_temp %>% dplyr::group_by(person) %>%
  dplyr::summarize(connection_conservative_firm = sum(connection_conservative_firm)) %>%
  dplyr::distinct()

# Merge that into core
tomerge <- dplyr::left_join(tomerge, firm_temp)
```

### Conservative membership
We create a composite variable identifying whether an individual was a member of any conservative organization
```{r}
# Calculate who has a 1 for either
memb_temp <- df_noemploy %>% dplyr::select(person, american_center_for_law_and_justice, 
                                           christian_legal_society_bar_association, 
                                           federalist_society, 
                                           heritage_foundation, 
                                           national_rifle_association_tie, 
                                           republican_national_lawyers_association_tie) %>%
                            dplyr::mutate(connection_conservative_group = ifelse(american_center_for_law_and_justice == 1 | 
                                            christian_legal_society_bar_association == 1 |
                                            federalist_society == 1 |
                                            heritage_foundation == 1 |
                                            national_rifle_association_tie == 1 |
                                            republican_national_lawyers_association_tie == 1,
                                            "1", "")) %>%
                           dplyr::select(person, connection_conservative_group) %>%
                           dplyr::distinct()

memb_temp$connection_conservative_group <- as.numeric(memb_temp$connection_conservative_group)

tomerge <- dplyr::left_join(tomerge, memb_temp)
```

### State senator/representative
There are 4 state senators/representatives that we code as conservative because they were registered members of the Republican party with well established ties
```{r}
employ_temp <- df %>% dplyr::select(person, employer_past) %>%
  dplyr::distinct() %>%
  dplyr::mutate(connection_conservative_gov = ifelse(employer_past == "Idaho State Senate" |
                                          employer_past == "Tennessee State Senate" |
                                          employer_past == "Georgia House of Representatives" |
                                          employer_past == "Oklahoma House of Representatives",
                                          "1", "")) %>%
  dplyr::select(person, connection_conservative_gov) %>%
  dplyr::distinct()

employ_temp$connection_conservative_gov <- as.numeric(employ_temp$connection_conservative_gov)

employ_temp <- employ_temp %>% dplyr::group_by(person) %>%
  dplyr::summarize(connection_conservative_gov = sum(connection_conservative_gov)) %>%
  dplyr::distinct()

tomerge <- dplyr::left_join(tomerge, employ_temp)
```

### Merge
```{r}
# Combine tomerge into a single column called conservative_agenda_connections
tomerge <- tomerge %>% dplyr::mutate(connection_conservative_all = ifelse(connection_conservative_lawschool == 1 |
                                                                                connection_conservative_firm == 1 |
                                                                                connection_conservative_group == 1 |
                                                                                connection_conservative_gov == 1,
                                                                              "1", "0")) %>%
                                                                    dplyr::distinct()

tomerge$connection_conservative_all <- as.numeric(tomerge$connection_conservative_all)
tomerge$connection_conservative_all[is.na(tomerge$connection_conservative_all)] <- 0

# Merge into master list
final <- dplyr::left_join(core, tomerge)

# Replace NAs with 0
final[c("connection_trump_all", "connection_trump_interviews", "connection_trump_campaign", "connection_trumpally_all", "connection_trumpally_innercircle", "connection_trumpally_us_senate", "connection_trumpally_state_gov", "connection_trumpally_rnc", "connection_trumpally_administration", "connection_conservative_all", "connection_conservative_lawschool", "connection_conservative_firm", "connection_conservative_group", "connection_conservative_gov", "connection_careerprofessional_all", "connection_other_all")][is.na(final[c("connection_trump_all", "connection_trump_interviews", "connection_trump_campaign", "connection_trumpally_all", "connection_trumpally_innercircle", "connection_trumpally_rnc", "connection_trumpally_us_senate", "connection_trumpally_state_gov", "connection_trumpally_administration", "connection_conservative_all", "connection_conservative_lawschool", "connection_conservative_firm", "connection_conservative_group", "connection_conservative_gov", "connection_careerprofessional_all", "connection_other_all")])] <- 0

# Fix edge column NA conversions
final$connection_trumpally_us_senate[is.na(final$connection_trumpally_us_senate)] <- 1
final$connection_trumpally_state_gov[is.na(final$connection_trumpally_state_gov)] <- 1
final$connection_trumpally_rnc[is.na(final$connection_trumpally_rnc)] <- 1

# Merge everything else into that now
df_noemploy <- dplyr::left_join(final, df_noemploy)

# Reorder columns
df_noemploy <- df_noemploy %>% dplyr::select("person", "position_current", "source",
                "date_of_appointment",
                "date_of_birth", "age", "birthplace", "sex",
                "barr_appointment", "political_affiliation", 
                "bachelors", "masters", "doctorate", "law", "law_conservative_score", "law_rank",
                "connection_trump_all", "connection_trump_interviews", "connection_trump_campaign",
                "connection_trumpally_all", "connection_trumpally_innercircle", "connection_trumpally_us_senate", "connection_trumpally_state_gov", "connection_trumpally_rnc",
                "connection_conservative_all", "connection_conservative_lawschool", "connection_conservative_firm", "connection_conservative_group", "connection_conservative_gov",
                "federalist_society", "heritage_foundation", "american_center_for_law_and_justice", "christian_legal_society_bar_association",  "national_rifle_association_tie", "republican_national_lawyers_association_tie",
                "connection_careerprofessional_all",
<<<<<<< HEAD
                "connection_other_all")
=======
                "connection_other_all",
                "american_center_for_law_and_justice", "christian_legal_society_bar_association", "federalist_society", "heritage_foundation", "national_rifle_association_tie", "republican_national_lawyers_association_tie")
<<<<<<< HEAD

# Fix NAs that should be 0's
df_noemploy[c("connection_trump_all", "connection_trump_interviews", "connection_trump_campaign", "connection_trumpally_all", "connection_trumpally_innercircle", "connection_trumpally_us_senate", "connection_trumpally_state_gov", "connection_trumpally_rnc", "connection_conservative_all", "connection_conservative_lawschool", "connection_conservative_firm", "connection_conservative_group", "connection_conservative_gov", "connection_careerprofessional_all", "connection_other_all")][is.na(df_noemploy[c("connection_trump_all", "connection_trump_interviews", "connection_trump_campaign", "connection_trumpally_all", "connection_trumpally_innercircle", "connection_trumpally_rnc", "connection_trumpally_us_senate", "connection_trumpally_state_gov", "connection_conservative_all", "connection_conservative_lawschool", "connection_conservative_firm", "connection_conservative_group", "connection_conservative_gov", "connection_careerprofessional_all", "connection_other_all")])] <- 0
=======
>>>>>>> b30f278ec2d9c801478220f5ab3cb1948a0bc43d
>>>>>>> 14b234190399f01bdd7011f398febe6bc4241f77
```

# Save
We have 2 versions of the data. One where the appointee is the unit of analysis and the other where prior employment for each appointtee is also listed, creating multiple rows per person.
```{r}
write.csv(df, paste0(here::here(), "/data/", "all_withemploy.csv"))
write.csv(df_noemploy, paste0(here::here(), "/data/", "all_noemploy.csv"))
```
