==> docs/04_Gannon_Corruption_Viz.Rmd <==
---
title: "04 Visualizations"
author: "J Andres Gannon, Danil Gromakov, Katelyn Hess, and Yulia Krylova"
date: "May 2020"
output:
  html_document:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(magrittr)
library(ggplot2)
```

This document produces some descriptive summary statistics of the new data collected regarding US prosecutor and Department of Justice officials appointed during the Trump administration

# Load data
We use the aggregated and combined data on both DOJ officials and prosecutors here
```{r}
df <- read.csv(file = paste0(here::here(), '/data/all_noemploy.csv'))
df_employ <- read.csv(file = paste0(here::here(), '/data/all_withemploy.csv'))
```

# Analyze data quality
## Missingness
We identify which columns are missing data. Much of the missingness is insignificant, since most personnel do not have doctorates or known membership in some of these societies. We do not convert these NAs to 0's because not finding information confirming their membership does not mean they were not members of these organizations
```{r, fig.width = 16, fig.height = 16}
visdat::vis_miss(df, sort_miss = TRUE)
visdat::vis_dat(df)
naniar::gg_miss_var(df)
```

## Covariate distributions
We can examine distributions of the numeric variables to identify outliers or problems in data quality. Everything looks largely correct
```{r, fig.width = 16, fig.height = 16}
DataExplorer::plot_qq(df)
DT::datatable(funModeling::df_status(df))
funModeling::plot_num(df)
SmartEDA::ExpNumViz(df)
```

# Law school rankings
Show difference in conservatism of law schools in the Trump admin vs all of them
## Load full law school data
```{r}
school <- read.csv(file = paste0(here::here(), '/data/ideology_lawschool.csv'))
school$score_scaled <- scale(school$conservatism)

school <- school %>% dplyr::select(-c("conservatism"))

df_law <- df %>% dplyr::select(person, law) %>%
  dplyr::distinct() %>%
  dplyr::rename(school = law) %>%
  dplyr::group_by(school) %>%
  dplyr::summarise(count = dplyr::n_distinct(person)) %>%
  dplyr::rename(persons = count)

df_law$trump <- 1

school_all <- dplyr::left_join(school, df_law)
school_all$school <- reorder(school_all$school, school_all$rank)
school_all$persons[is.na(school_all$persons)] <- 0
school_all$trump <- as.factor(school_all$trump)
school_all$labels <- school_all$school
school_all$labels[school_all$trump == 0] <- NA
school_all$rank[school_all$school == "Harvard University"] <- 2
school_all$trump <- factor(school_all$trump, levels=c(levels(school_all$trump), 0))
school_all$trump[is.na(school_all$trump)] <- 0

gd <- school_all %>% dplyr::group_by(trump) %>%
  dplyr::summarise(avg_rank = mean(rank, na.rm = TRUE),
                   avg_score = mean(score_scaled, na.rm = TRUE))

ggplot2::ggplot(school_all, aes(x = reorder(school, dplyr::desc(school)) , y = score_scaled, color = trump)) + 
  geom_point() + 
  coord_flip() + 
  scale_x_discrete(breaks = c("Harvard University", "Whittier College"), labels = c("1", "200")) +
  scale_y_continuous(breaks = c(-2, 0, 2), labels = c("Liberal", "Neutral", "Conservative")) +
  scale_color_discrete(limits = c("1", "0"), labels = c("Trump appointee", "Not Trump appointee")) +
  theme_light() +
  theme(text = element_text(size = 14),
        legend.position = "bottom", legend.direction = "horizontal", legend.title = element_blank()) + 
  labs(title = "Law School Ideology Bias in Trump Prosecutors", x = "Law Schools in Rank Order", y = "Liberal-Conservative Scale")
```

# Membership overlaps
```{r}
# Trump ties
upset_trump <- df %>% dplyr::select(dplyr::starts_with("connection_trump_")) %>%
  dplyr::select(-c("connection_trump_all")) %>%
  dplyr::rename("Trump personal interview" = connection_trump_interviews, "Trump campaign connection" = connection_trump_campaign)

UpSetR::upset(upset_trump,
      number.angles = 0, point.size = 3.5, line.size = 2, text.scale = 2,
      mainbar.y.label = "Officials", order.by = "freq")

<<<<<<< HEAD
=======
# Trump ally ties
upset_trumpally <- df %>% dplyr::select(dplyr::starts_with("connection_trumpally_")) %>%
  dplyr::select(-c("connection_trumpally_all")) %>%
  dplyr::rename("Trump ally inner circle" = connection_trumpally_innercircle, "Trump ally, US Senate" = connection_trumpally_us_senate, "Trump ally, state gov" = connection_trumpally_state_gov, "Trump ally, RNC" = connection_trumpally_rnc)

UpSetR::upset(upset_trumpally,
      number.angles = 0, point.size = 3.5, line.size = 2, text.scale = 2,
      mainbar.y.label = "Officials", order.by = "freq")

>>>>>>> b30f278ec2d9c801478220f5ab3cb1948a0bc43d
# Conservative group ties
upset_conservative <- df %>% dplyr::select(dplyr::starts_with("connection_conservative_")) %>%
  dplyr::select(-c("connection_conservative_all")) %>%
  dplyr::rename("Conservative law school" = connection_conservative_lawschool, "Conservative law firm" = connection_conservative_firm, "Conservative society membership" = connection_conservative_group, "Conservative state official" = connection_conservative_gov)

UpSetR::upset(upset_conservative,
      number.angles = 0, point.size = 3.5, line.size = 2, text.scale = 2,
      mainbar.y.label = "Officials", order.by = "freq")

# All ties
upset_all <- df %>% dplyr::select(dplyr::ends_with("_all")) %>%
<<<<<<< HEAD
  dplyr::rename("Trump direct" = connection_trump_all, "Trump allies" = connection_trumpally_all, "Conservative background" = connection_conservative_all, "Carrer Professional" = connection_careerprofessional_all, "Other" = connection_other_all)
=======
  dplyr::rename("Trump connections" = connection_trump_all, "Trump allies" = connection_trumpally_all, "Conservative background" = connection_conservative_all, "Carrer Professional" = connection_careerprofessional_all, "Other" = connection_other_all)
>>>>>>> b30f278ec2d9c801478220f5ab3cb1948a0bc43d

UpSetR::upset(upset_all,
      number.angles = 0, point.size = 3.5, line.size = 2, text.scale = 2,
      mainbar.y.label = "Officials", order.by = "freq")
```

# Distributions
We can plot the percent distribution of persons with associations in conservative law schools, firms, and organizations
```{r}
# Conservative bias
conservative <- df %>% dplyr::select(dplyr::starts_with("connection_conservative")) %>%
  dplyr::select(-c("connection_conservative_all", "connection_conservative_gov"))
sumdata <- data.frame(value = apply(conservative, 2, sum))
sumdata$key <- rownames(sumdata)

ggplot(sumdata, aes(x = key, y = value, fill = key)) +
  geom_bar(stat = "identity") +
  scale_x_discrete(labels = c("Conservative Law Firm", "Conservative Membership", "Conservative Law School")) +
  labs(title = "Number of Trump Officials with Questionable Histories", x = "History", y = "Number of Officials") +
  theme(plot.title = element_text(hjust = 0.5),panel.grid = element_blank(),
        text = element_text(size = 14),
        axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1),
        legend.position = "none")

# Low ranking law schools
ggplot2::ggplot(df, aes(x = law_rank, y = ..count..)) +
  geom_histogram() +
  labs(title = "Trump Officials and Law School Ranking", x = "Law School Ranking", y = "Number of Officials") +
  scale_x_continuous(breaks = c(1, 50, 100, 150, 200)) +
  theme(plot.title = element_text(hjust = 0.5),panel.grid = element_blank(),
        text = element_text(size = 14))
```
